Removed Errors Part 1

+++ include/input.h
@@ -3,6 +3,6 @@
 
 //----------Function Prototypes---------
 int getInteger(char *msg ,int min, int max);
-float getFloating(char *msg ,int min, int max);
+float getFloating(char *msg ,float min,float max);
 
 #endif
